Fix issue by switching to double floating point precision

+++ GravSim/particleRasterizer.cpp
@@ -22,25 +22,25 @@ void particleRasterizer::initRast_A(RastInit details) {
 
 	createBuffers();
 
-	(*planets)[0].pos_0 = glm::vec3(0.0f, 0.0f, 0.0f);
-	(*planets)[0].radius = 6378e3f;
-	(*planets)[0].vel_0 = glm::vec3(0.0f, 0.0f, 0.0f);
-	(*planets)[0].mass = 5.9722e24f;
-	(*planets)[0].axis = glm::vec3(0.0f, glm::asin(glm::radians(23.5f)), glm::acos(glm::radians(23.5f)));
+	(*planets)[0].pos_0 = glm::dvec3(0.0f, 0.0f, 0.0f);
+	(*planets)[0].radius = 6378e3;
+	(*planets)[0].vel_0 = glm::dvec3(0.0f, 0.0f, 0.0f);
+	(*planets)[0].mass = 5.9722e24;
+	(*planets)[0].axis = glm::dvec3(0.0f, glm::asin(glm::radians(23.5)), glm::acos(glm::radians(23.5)));
 	(*planets)[0].theta = 0.0f;
 
-	(*planets)[1].pos_0 = glm::vec3(0.4055e9/*10.0f*/, 0.0f, 0.0f);
-	(*planets)[1].radius = 1738e3f/*0.5f*/;
-	(*planets)[1].vel_0 = glm::vec3(0.0f, 0.970e3/*0.5f*/, 0.0f);
-	(*planets)[1].mass = 0.07346e24f;
-	(*planets)[1].axis = glm::vec3(0.0f, 0.0f, 1.0f);
+	(*planets)[1].pos_0 = glm::dvec3(0.4055e9/*10.0f*/, 0.0f, 0.0f);
+	(*planets)[1].radius = 1738e3/*0.5f*/;
+	(*planets)[1].vel_0 = glm::dvec3(0.0f, 0.970e3/*0.5f*/, 0.0f);
+	(*planets)[1].mass = 0.07346e24;
+	(*planets)[1].axis = glm::dvec3(0.0f, 0.0f, 1.0f);
 	(*planets)[1].theta = 0.0f;
 
 	for (uint32_t i = 2; i < planets->size(); i++) {
 		(*planets)[i].pos_0 = glm::vec3(glm::linearRand<float>(0, 10) * 7e7, glm::linearRand<float>(0, 10) * 8e7, glm::linearRand<float>(0, 10) * 1e1);
 		(*planets)[i].radius = glm::linearRand<float>(0.1f, 2.0f)*1e6f;
 		//(*planets)[i].vel = glm::vec3(glm::linearRand<float>(0, 1)*1e3, glm::linearRand<float>(0, 1)*1e3, glm::linearRand<float>(0, 1)*1e0 );
-		(*planets)[i].vel_0 = glm::sqrt(float(6.67e-11) * (*planets)[0].mass * glm::linearRand<float>(0.7f, 1.4f) / glm::length((*planets)[i].pos_0)) * glm::cross(glm::normalize((*planets)[i].pos_0), glm::vec3(0.0f, 0.0f, 1.0f));
+		(*planets)[i].vel_0 = glm::sqrt(float(6.67e-11) * (*planets)[0].mass * glm::linearRand<float>(0.7f, 1.4f) / glm::length((*planets)[i].pos_0)) * glm::cross(glm::normalize((*planets)[i].pos_0), glm::dvec3(0.0f, 0.0f, 1.0f));
 		//(*planets)[i].mass = glm::linearRand<float>(0.1f, 2.0f);
 		(*planets)[i].mass = (*planets)[0].mass * glm::pow((*planets)[i].radius / (*planets)[0].radius, 3.0f);
 		(*planets)[i].axis = glm::vec3(glm::linearRand<float>(0, 1), glm::linearRand<float>(0, 1), glm::linearRand<float>(0, 1));
